requirements.txt
pytesseract>=0.3.10
pdf2image>=1.16.3
pdfplumber>=0.10.3
Pillow>=10.2.0
opencv-python-headless>=4.9.0
python-docx>=1.1.0
anthropic>=0.40.0
httpx>=0.27.0
pytest>=8.2.0
python-dotenv>=1.0.1
